+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/Golden.hs
@@ -9,24 +9,38 @@ module Test.Cardano.Ledger.Binary.Golden (
   expectDecoderSuccessAnnWith,
   expectDecoderFailureAnn,
   expectDecoderResultOn,
+  toPackageGolden,
+  goldenForToCBOR,
+  goldenForEncCBOR,
+  cborGoldenSpec,
+  cborAnnGoldenSpec,
 ) where
 
 import Cardano.Ledger.Binary (
   Annotator,
   DecCBOR (..),
   DecoderError,
+  EncCBOR (..),
   ToCBOR (..),
   Version,
   decodeFullAnnotator,
+  serialize,
   toLazyByteString,
  )
 import qualified Cardano.Ledger.Binary as Binary
+import qualified Cardano.Ledger.Binary.Plain as Plain
+import qualified Data.ByteString.Lazy as BSL
 import Data.TreeDiff (ToExpr (..))
-import Data.Typeable (Proxy (..))
-import GHC.Stack (HasCallStack)
+import Data.Typeable (Proxy (..), typeOf)
 import Test.Cardano.Ledger.Binary.Plain.Golden (Enc)
-import Test.Cardano.Ledger.Binary.RoundTrip (embedTripAnnExpectation)
-import Test.Hspec (Expectation, expectationFailure, shouldBe)
+import Test.Cardano.Ledger.Binary.RoundTrip (
+  embedTripAnnExpectation,
+  roundTripAnnRangeExpectation,
+  roundTripCborRangeExpectation,
+ )
+import Test.Cardano.Ledger.Binary.TreeDiff (CBORBytes (..))
+import Test.Hspec
+import qualified Test.Hspec.Golden as Golden (Golden (..))
 
 decodeEnc :: forall a. DecCBOR (Annotator a) => Version -> Enc -> Either DecoderError a
 decodeEnc version enc = decodeFullAnnotator @a version (Binary.label $ Proxy @(Annotator a)) decCBOR bytes
@@ -88,3 +102,104 @@ expectDecoderResultOn version enc expected f =
     version
     (\x _ -> f x `shouldBe` f expected)
     enc
+
+toPackageGolden :: (FilePath -> IO FilePath) -> Golden.Golden g -> IO (Golden.Golden g)
+toPackageGolden mkFullPath g = do
+  fullPathGoldenFile <- mkFullPath $ Golden.goldenFile g
+  fullPathActualFile <- mapM mkFullPath $ Golden.actualFile g
+  pure $
+    g
+      { Golden.goldenFile = fullPathGoldenFile
+      , Golden.actualFile = fullPathActualFile
+      }
+
+-- | `Golden` specification for `ToCBOR`
+goldenForToCBOR ::
+  ToCBOR a =>
+  -- | Path to the golden file relative to the root of the package
+  FilePath ->
+  -- | Value, which in an encoded form will be expected to produce the same contents as in the
+  -- golden file.
+  a ->
+  Golden.Golden BSL.ByteString
+goldenForToCBOR goldenFileName t =
+  Golden.Golden
+    { Golden.output = Plain.serialize t
+    , Golden.encodePretty = show . CBORBytes . BSL.toStrict
+    , Golden.writeToFile = BSL.writeFile
+    , Golden.readFromFile = BSL.readFile
+    , Golden.goldenFile = goldenFileName
+    , Golden.actualFile = Nothing
+    , Golden.failFirstTime = False
+    }
+
+-- | `Golden` specification for `EncCBOR`
+goldenForEncCBOR ::
+  EncCBOR a =>
+  -- | Path to the golden file relative to the package
+  FilePath ->
+  -- | Protocol version to be used for encoding
+  Version ->
+  -- | Value, which in an encoded form will be expected to produce the same contents as in the
+  -- golden file.
+  a ->
+  Golden.Golden BSL.ByteString
+goldenForEncCBOR goldenFileName version t =
+  Golden.Golden
+    { Golden.output = serialize version t
+    , Golden.encodePretty = show . CBORBytes . BSL.toStrict
+    , Golden.writeToFile = BSL.writeFile
+    , Golden.readFromFile = BSL.readFile
+    , Golden.goldenFile = goldenFileName
+    , Golden.actualFile = Nothing
+    , Golden.failFirstTime = False
+    }
+
+-- | Check `EncCBOR` golden spec as well as roundtripping of the golden example with `DecCBOR`
+cborGoldenSpec ::
+  forall a.
+  ( Eq a
+  , Show a
+  , EncCBOR a
+  , DecCBOR a
+  , HasCallStack
+  ) =>
+  -- | Action to get the full path, usually will be @Paths_<package_name>.getDataFileName@
+  (FilePath -> IO FilePath) ->
+  -- | File path to the golden file relative to the root of the package
+  FilePath ->
+  -- | Protocol version to be used for encoding
+  Version ->
+  -- | Value, which in an encoded form will be expected to produce the same contents as in the
+  -- golden file.
+  a ->
+  Spec
+cborGoldenSpec mkFullPath goldenFileName version a = do
+  describe (show (typeOf a)) $ do
+    it "Golden" $ toPackageGolden mkFullPath $ goldenForEncCBOR goldenFileName version a
+    it "RoundTrip Golden Example" $ roundTripCborRangeExpectation version version a
+
+-- | Check `ToCBOR` golden spec as well as roundtripping of the golden example with `DecCBOR` for
+-- its `Annotator` version.
+cborAnnGoldenSpec ::
+  forall a.
+  ( Eq a
+  , Show a
+  , ToCBOR a
+  , DecCBOR (Annotator a)
+  , HasCallStack
+  ) =>
+  -- | Action to get the full path, usually will be @Paths_<package_name>.getDataFileName@
+  (FilePath -> IO FilePath) ->
+  -- | File path to the golden file relative to the root of the package
+  FilePath ->
+  -- | Protocol version to be used for decoding
+  Version ->
+  -- | Value, which in an encoded form will be expected to produce the same contents as in the
+  -- golden file.
+  a ->
+  Spec
+cborAnnGoldenSpec mkFullPath goldenFileName version a = do
+  describe (show (typeOf a)) $ do
+    it "Golden" $ toPackageGolden mkFullPath $ goldenForToCBOR goldenFileName a
+    it "RoundTrip Golden Example" $ roundTripAnnRangeExpectation version version a